Add user getUser and getFriends functions to the controller, and refactored functions to modulate DB interactions in functions living in the model
